We should remove the lock in the Session.HasResponder. This was an overlocking which was causing deadlock when sending and receiving larget amount FIX messages simultaneously.
A detailed description of the issue can be found here:
http://lists.quickfixn.com/pipermail/quickfixn-quickfixn.com/2012q2/000202.html

Removing the lock doesn't affect the current unit tests and acceptance tests.
This issue was hard to be reproduced through unit tests as it requires a TCP
connection between a server and a client. I have attched two additional projects to verify the correctness of the change.
Launch the FIXServer first and then FIXClient. Following the instructions to
issue 10,000 orders from the FIXClient, wihle sending the FIXServer will start
sending execution reports back.


Former-commit-id: dc0d5c3a91c3b53f2b3e652a84da3c65f48a333d [formerly f630016839493d93600a7c79994021272a2305ea] [formerly c8646c01fa2e03654dd6bdbc7ca52be8af079d37 [formerly fd2487b72cd9ac36e54eb7154b21a510d5037732]]
Former-commit-id: 3052ae5debdb6d9a88e29a1a1713c723ae54066c [formerly f6c04a44bd89c5f37f4f883f62df84d300ed1048]
Former-commit-id: f54e8ce923c41ee48942d90d796fc401087d0df5